fix(iconswitch): apply changes from PR review
